acpi: reorganize DSDT and SSDT table generation
The SSDT tables extend the DSDT, so intially it made sense to keep them
close together. But a system can have multiple SSDTs that may be better
to define in a separate module.

The existing SSDT is also likely to be removed in the future, so keeping
its components isolate makese it easier to delete it.
