+++ lib/propolis/src/firmware/acpi/dsdt.rs
@@ -42,17 +42,6 @@ fn dsdt_sdt_edk2_style() -> Sdt {
     Sdt::new(*b"DSDT", 36, 1, *b"INTEL ", *b"OVMF    ", 0x4)
 }
 
-fn ssdt_sdt_edk2_style() -> Sdt {
-    // For SSDTs, the OEM table ID needs to be different for each table. Refer
-    // to ACPI rev. 6.6 section 5.2.11.2 "Secondary System Description Table
-    // (SSDT)" for more information.
-    //
-    // The SSDT provided by EDK2 can also be removed entirely in the future
-    // because it only holds unsupported sleep states (S2 and S3) and FWDT
-    // memory region.
-    Sdt::new(*b"SSDT", 36, 1, *b"REDHAT", *b"OVMF    ", 0x1)
-}
-
 /// The ACPI scope in which DsdtGenerators are placed.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DsdtScope {
@@ -110,12 +99,7 @@ impl<'a> Aml for DsdtGeneratorAml<'a> {
 /// Currently, only the S0->S5 transition is handled explicitly, but S0->S0 is
 /// also handled properly as a no-op, and so the value of 5 is never checked
 /// directly.
-///
-/// Transitions to S3 and S4 are inherited from the original EDK2 tables and
-/// should probably be removed in the future.
 const PM1A_CNT_SLP_TYP_S0: u8 = 5;
-const PM1A_CNT_SLP_TYP_S3: u8 = 1;
-const PM1A_CNT_SLP_TYP_S4: u8 = 2;
 const PM1A_CNT_SLP_TYP_S5: u8 = 0;
 
 pub struct DsdtConfig<'a> {
@@ -176,33 +160,6 @@ impl<'a> Aml for Dsdt<'a> {
         let mut sdt = dsdt_sdt_edk2_style();
         sdt.append_slice(dsdt.as_slice());
         sdt.to_aml_bytes(sink);
-    }
-}
-
-/// Describe a sleep state.
-struct SleepState<'a> {
-    state: &'a str,
-    pm1a_cnt: u8,
-}
-
-impl<'a> SleepState<'a> {
-    fn new(state: &'a str, pm1a_cnt: u8) -> Self {
-        Self { state, pm1a_cnt }
-    }
-}
-
-impl<'a> Aml for SleepState<'a> {
-    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
-        aml::Name::new(
-            self.state.into(),
-            &aml::Package::new(vec![
-                &self.pm1a_cnt, // PM1a_CNT.SLP_TYP value to enter this sleep state.
-                &aml::ZERO, // PM1b_CNT.SLP_TYP value. PM1b is not currently used in Propolis.
-                &aml::ZERO, // Reserved.
-                &aml::ZERO, // Reserved.
-            ]),
-        )
-        .to_aml_bytes(sink);
     }
 }
 
@@ -499,14 +456,6 @@ impl Aml for PciRootBridgeCrs {
     }
 }
 
-fn aml_len(vec: &[&dyn Aml]) -> usize {
-    let mut sink = Vec::new();
-    for aml in vec {
-        aml.to_aml_bytes(&mut sink);
-    }
-    sink.len()
-}
-
 /// Number of devices in the PCI0 root bridge to link to the interrupt
 /// controller.
 ///
@@ -555,7 +504,7 @@ impl Aml for PciRootBridgePrt {
                     sources[idx as usize]
                 };
 
-                ptr_entries.push(PrtEntry { device, pin, source });
+                ptr_entries.push(PrtEntry::new(device, pin, source));
             }
         }
         let ptr = ptr_entries.iter().map(|p| p as &dyn Aml).collect();
@@ -565,32 +514,6 @@ impl Aml for PciRootBridgePrt {
             NOT_SERIALIZED,
             vec![&aml::Return::new(&aml::Package::new(ptr))],
         )
-        .to_aml_bytes(sink);
-    }
-}
-
-/// Low-word of an _ADR value that refers to all PCI functions.
-///
-/// <https://uefi.org/htmlspecs/ACPI_Spec_6_4_html/06_Device_Configuration/Device_Configuration.html#adr-object-address-encodings>
-const PCI_ADR_ALL_FUNC: u32 = 0xffff;
-
-/// Representation of an entry in the _PRT table.
-struct PrtEntry<'a> {
-    device: u8,
-    pin: u8,
-    source: &'a str,
-}
-
-impl<'a> Aml for PrtEntry<'a> {
-    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
-        let addr: aml::DWord = ((self.device as u32) << 16) | PCI_ADR_ALL_FUNC;
-
-        aml::Package::new(vec![
-            &addr,
-            &self.pin,
-            &aml::Path::new(self.source),
-            &aml::ZERO,
-        ])
         .to_aml_bytes(sink);
     }
 }
@@ -931,48 +854,116 @@ impl<'a> Aml for PciRootBridgeLpc<'a> {
     }
 }
 
-/// Represents a PCI IRQ link in the LPC device.
-struct Lnk<'a> {
-    name: &'a str,
-    uid: u32,
+/// Helper function to compute the length of a vector of Amls. Useful when
+/// computing the offset of objects inside a scope.
+pub fn aml_len(vec: &[&dyn Aml]) -> usize {
+    let mut sink = Vec::new();
+    for aml in vec {
+        aml.to_aml_bytes(&mut sink);
+    }
+    sink.len()
 }
 
-impl<'a> Lnk<'a> {
-    fn new(name: &'a str, uid: u32) -> Self {
-        Self { name, uid }
+/// Describe a sleep state.
+///
+/// ACPI rev. 6.6 section 7.4.2 "\_Sx (System States)"
+pub struct SleepState<'a> {
+    state: &'a str,
+    pm1a_cnt: u8,
+}
+impl<'a> SleepState<'a> {
+    pub fn new(state: &'a str, pm1a_cnt: u8) -> Self {
+        Self { state, pm1a_cnt }
     }
 }
-
-impl<'a> Lnk<'a> {
-    // AML code for the special SCI link.
-    // <https://github.com/oxidecomputer/edk2/blob/f33871f488bfbbc080e0f7e3881e04d0db0b6367/OvmfPkg/AcpiTables/Dsdt.asl#L200-L220>
-    fn to_aml_bytes_sci(&self, sink: &mut dyn AmlSink) {
-        aml::Device::new(
-            "LNKS".into(),
-            vec![
-                &names::hid(&aml::EISAName::new(devids::PCI_INT_LINK)),
-                &names::uid(&aml::ZERO),
-                &names::sta(&0x0b_u64),
-                &methods::srs(1, NOT_SERIALIZED, vec![]),
-                &methods::dis(0, NOT_SERIALIZED, vec![]),
-                &names::prs(&aml::ResourceTemplate::new(vec![
-                    &aml::Interrupt::new(true, false, false, true, vec![0x09]),
-                ])),
-                &methods::crs(
-                    0,
-                    NOT_SERIALIZED,
-                    vec![&aml::Return::new(&paths::prs())],
-                ),
-            ],
+impl<'a> Aml for SleepState<'a> {
+    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
+        aml::Name::new(
+            self.state.into(),
+            &aml::Package::new(vec![
+                &self.pm1a_cnt, // PM1a_CNT.SLP_TYP value to enter this sleep state.
+                &aml::ZERO, // PM1b_CNT.SLP_TYP value. PM1b is not currently used in Propolis.
+                &aml::ZERO, // Reserved.
+                &aml::ZERO, // Reserved.
+            ]),
         )
         .to_aml_bytes(sink);
     }
 }
 
+/// Describe an entry of the _PRT table.
+///
+/// ACPI rev. 6.6 section 6.2.14 "_PRT (PCI Routing Table)"
+pub struct PrtEntry<'a> {
+    device: u8,
+    pin: u8,
+    source: &'a str,
+}
+impl<'a> PrtEntry<'a> {
+    pub fn new(device: u8, pin: u8, source: &'a str) -> Self {
+        Self { device, pin, source }
+    }
+}
+impl<'a> Aml for PrtEntry<'a> {
+    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
+        // PCI device address format: <device number><device function>
+        // Device function 0xffff represents all functions on a device.
+        //
+        // ACPI rev. 6.6 table 6.2 "ADR Object Address Encodings"
+        let addr: aml::DWord = ((self.device as u32) << 16) | 0xffff;
+
+        aml::Package::new(vec![
+            &addr,
+            &self.pin,
+            &aml::Path::new(self.source),
+            &aml::ZERO,
+        ])
+        .to_aml_bytes(sink);
+    }
+}
+
+/// Describe a PCI IRQ link device of the LPC.
+///
+/// ACPI rev. 6.6 section 6.2.14.1 "Example: Using _PRT to Describe PCI IRQ Routing"
+struct Lnk<'a> {
+    name: &'a str,
+    uid: u32,
+}
+impl<'a> Lnk<'a> {
+    fn new(name: &'a str, uid: u32) -> Self {
+        Self { name, uid }
+    }
+}
 impl<'a> Aml for Lnk<'a> {
     fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
         if self.uid == 0 {
-            self.to_aml_bytes_sci(sink);
+            // AML code for the special SCI link.
+            // https://github.com/oxidecomputer/edk2/blob/f33871f488bfbbc080e0f7e3881e04d0db0b6367/OvmfPkg/AcpiTables/Dsdt.asl#L200-L220
+            aml::Device::new(
+                "LNKS".into(),
+                vec![
+                    &names::hid(&aml::EISAName::new(devids::PCI_INT_LINK)),
+                    &names::uid(&aml::ZERO),
+                    &names::sta(&0x0b_u64),
+                    &methods::srs(1, NOT_SERIALIZED, vec![]),
+                    &methods::dis(0, NOT_SERIALIZED, vec![]),
+                    &names::prs(&aml::ResourceTemplate::new(vec![
+                        &aml::Interrupt::new(
+                            true,
+                            false,
+                            false,
+                            true,
+                            vec![0x09],
+                        ),
+                    ])),
+                    &methods::crs(
+                        0,
+                        NOT_SERIALIZED,
+                        vec![&aml::Return::new(&paths::prs())],
+                    ),
+                ],
+            )
+            .to_aml_bytes(sink);
             return;
         }
 
@@ -1030,153 +1021,6 @@ impl<'a> Aml for Lnk<'a> {
             ],
         )
         .to_aml_bytes(sink);
-    }
-}
-
-/// Length in bytes of the SSDT header. Used to calculate the offset of other
-/// fields.
-///
-/// <https://uefi.org/htmlspecs/ACPI_Spec_6_4_html/05_ACPI_Software_Programming_Model/ACPI_Software_Programming_Model.html#secondary-system-description-table-fields-ssdt>
-const SSDT_HEADER_LEN: usize = 36;
-
-/// Byte offset of the FWDT OperationRegion offset address field in the SSDT
-/// table. This filed is updated in fwcfg.rs during table generation.
-///
-/// SSDT header (36 bytes) + External operation prefix (1 byte) +
-/// OperationRegion prefix (1 byte) + OperationRegion name (4 bytes) +
-/// OperationRegion space (1 byte) + DWordPrefix (1 byte)
-///
-/// <https://uefi.org/htmlspecs/ACPI_Spec_6_4_html/20_AML_Specification/AML_Specification.html#defopregion>
-pub const SSDT_FWDT_ADDR_OFFSET: usize = SSDT_HEADER_LEN + 8;
-
-/// Number of bytes used to store the offset address value in the FWDT
-/// OperationRegion. Size of a DWord.
-pub const SSDT_FWDT_ADDR_LEN: usize = 4;
-
-/// The SSDT table is an extension to DSDT table and can be used to extend
-/// resources defined in the DSDT.
-///
-/// <https://uefi.org/htmlspecs/ACPI_Spec_6_4_html/05_ACPI_Software_Programming_Model/ACPI_Software_Programming_Model.html#secondary-system-description-table-ssdt>
-pub struct Ssdt {
-    /// Offset of the area reserved for the FWDT OperationRegion data in the
-    /// overall ACPI tables storage.
-    fwdt_offset: usize,
-}
-
-impl Ssdt {
-    pub fn new(fwdt_offset: usize) -> Self {
-        Self { fwdt_offset }
-    }
-}
-
-// This implementation follows the original static EDK2 tables.
-//
-// https://github.com/oxidecomputer/edk2/blob/f33871f488bfbbc080e0f7e3881e04d0db0b6367/OvmfPkg/AcpiPlatformDxe/Qemu.c#L426-L466
-//
-// It can probably be further simplified or eliminated entirely in the future.
-impl Aml for Ssdt {
-    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
-        let mut ssdt = Vec::new();
-
-        // The FWDT OperationRegion is used to pass dynamic information about
-        // the instance from the platform to the virtual machine. The main
-        // information provided are the 32-bit and 64-bit PCI MMIO ranges.
-        //
-        // On boot, the \_SB.PCI0._CRS method reads FWDT and adjusts the PCI
-        // bus configuration based on the data it holds.
-        //
-        // This process can be removed if \_SB.PCI0._CRS is modified to return
-        // a static ResourceTemplate that is already populated with the right
-        // VM data.
-        aml::OpRegion::new(
-            "FWDT".into(),
-            aml::OpRegionSpace::SystemMemory,
-            &DWord::new(self.fwdt_offset as u32),
-            &DWord::new(0x30),
-        )
-        .to_aml_bytes(&mut ssdt);
-
-        // Sleep states.
-        //
-        // These sleep states are kept for consistency with the original static
-        // EDK2 tables. Propolis doesn't handle these state properly, so they
-        // should be removed in the future.
-        //
-        // These values don't use the SleepState struct to keep the generated
-        // AML code the same as the original EDK tables.
-        aml::Name::new(
-            "\\_S3_".into(),
-            &aml::Package::new(vec![
-                &Byte::new(PM1A_CNT_SLP_TYP_S3),
-                &Byte::new(0),
-                &Byte::new(0),
-                &Byte::new(0),
-            ]),
-        )
-        .to_aml_bytes(&mut ssdt);
-
-        aml::Name::new(
-            "\\_S4_".into(),
-            &aml::Package::new(vec![
-                &Byte::new(PM1A_CNT_SLP_TYP_S4),
-                &Byte::new(0),
-                &Byte::new(0),
-                &Byte::new(0),
-            ]),
-        )
-        .to_aml_bytes(&mut ssdt);
-
-        let mut sdt = ssdt_sdt_edk2_style();
-        sdt.append_slice(ssdt.as_slice());
-        sdt.to_aml_bytes(sink);
-    }
-}
-
-// Provides consistent DWord AML values. The acpi_tables crate minimizes
-// integers to the smallest word size that the number fits, but the original
-// EDK2-defined ACPI tables used wider-than-necessary integers in some places.
-//
-// We retained this quirk to avoid unnecessary differences in ACPI tables when
-// having Propolis generate them, but future ACPI table versions have no
-// particular need to keep this.
-struct DWord {
-    value: u32,
-}
-
-impl DWord {
-    fn new(value: u32) -> Self {
-        Self { value }
-    }
-}
-
-impl Aml for DWord {
-    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
-        sink.byte(0x0c); // DWordPrefix
-        sink.dword(self.value);
-    }
-}
-
-// Provides consistent Byte AML values. The acpi_tables crate minimizes
-// integers to the smallest word size that the number fits, but the original
-// EDK2-defined ACPI tables used wider-than-necessary integers in some places.
-//
-// We retained this quirk to avoid unnecessary differences in ACPI tables when
-// having Propolis generate them, but future ACPI table versions have no
-// particular need to keep this.
-struct Byte {
-    value: u8,
-}
-
-impl Byte {
-    fn new(value: u8) -> Self {
-        Self { value }
-    }
-}
-
-impl Aml for Byte {
-    fn to_aml_bytes(&self, sink: &mut dyn AmlSink) {
-        sink.byte(0x0a); // BytePrefix
-        sink.byte(self.value);
     }
 }
 
@@ -1391,13 +1235,5 @@ mod test {
         );
 
         sink.clear();
-
-        // Validate FWDT offset address field offset.
-        Ssdt::new(0xabc).to_aml_bytes(&mut sink);
-        assert_eq!(
-            sink[SSDT_FWDT_ADDR_OFFSET
-                ..(SSDT_FWDT_ADDR_OFFSET + SSDT_FWDT_ADDR_LEN)],
-            0xabc_u32.to_le_bytes()
-        );
     }
 }
